+++ yadmin/yamin-application/yadmin-application-chat/src/main/java/com/ytrue/yadmin/service/ChatContactService.java
@@ -20,7 +20,7 @@ public interface ChatContactService extends IService<ChatContact> {
      * @param contactId
      * @return
      */
-    List<ContactVO> getMyContactById(Integer contactId);
+    List<ContactVO> getMyContactById(Long contactId);
 
     /**
      * 获得当前用户侧栏聊天消息列表，限制是前150条
@@ -28,5 +28,6 @@ public interface ChatContactService extends IService<ChatContact> {
      * @param contactId
      * @return
      */
-    List<ContactVO> getMySidebarMessageById(Integer contactId);
+    List<ContactVO> getMySidebarMessageById(Long contactId);
+
 }
